+++ Changes.md
@@ -5,6 +5,7 @@ Version 4.0.0.1 - TBR
 -----------------------------
   1. Refactor - Moodle 4.0 version to cope with course format API changes.
   2. Fix 'Error during course creation if defaultdisplayblocks is empty' - CONTRIB-8824.
+  3. Add the ability to specify icon font classes for the close and open toggle icons.
 
 Version 3.11.0.5 - 15/11/21
 -----------------------------